==> notebooks/00-download-fivethirtyeight-data.ipynb <==
# Download Forecast Data From *FiveThirtyEight*

This notebook downloads the current predictions for the House and Senate races from *FiveThirtyEight* at a given moment and puts the resulting CSV in the `data/fivethirtyeight` directory.

import requests
import pandas as pd
import json

## House Forecast

The data we want is embedded in a JSON file on the *FiveThirtyEight* site that populates the webpage.

HOUSE_URL = "https://projects.fivethirtyeight.com/2018-midterm-election-forecast/house/home.json"
house_data = requests.get(HOUSE_URL).json()

## Extract Candidate Win Probabilities

house_candidates = []
for d in house_data["districtForecasts"]:
    for f in d["forecast"]:
        if d["state"] != "US":
            candidate_dict = {
                "state": d["state"],
                "district": d["district"]        
            }
            candidate_dict["candidate"] = f["candidate"]
            candidate_dict["party"] = f["party"]
            candidate_dict["classic_prob"] = f["models"]["classic"]["winprob"]
            house_candidates.append(candidate_dict)
            
house_candidates = pd.DataFrame(house_candidates)
house_candidates.head()

len(house_candidates)

house_candidates["state"].nunique()

house_candidates.to_csv(
    "../data/fivethirtyeight/house_candidate_odds.csv",
    index = None
)

## Senate Forecast

SENATE_URL = "https://projects.fivethirtyeight.com/2018-midterm-election-forecast/senate/home.json"

senate_data = requests.get(SENATE_URL).json()

### General forecast

senate_candidates = []
for d in senate_data["seatForecasts"]:
    for f in d["forecast"]:
        if d["state"] != "US":
            senate_candidate_dict = {
                "state": d["state"],
                "class": d["class"]
            }
            senate_candidate_dict["candidate"] = f["candidate"]
            senate_candidate_dict["party"] = f["party"]
            senate_candidate_dict["classic_prob"] = f["models"]["classic"]["winprob"]
            senate_candidates.append(senate_candidate_dict)
            
senate_candidates = pd.DataFrame(senate_candidates)

senate_candidates.head()

len(senate_candidates)

senate_candidates["state"].nunique()

senate_candidates.to_csv(
    "../data/fivethirtyeight/senate_candidate_odds.csv",
    index = None
)

### Voter Power Index

senate_seats = []
for senate_seat in senate_data["seatForecasts"]:
    if "vpi" in senate_seat.keys():
        senate_seat_dict = {
            "state": senate_seat["state"],
            "vpi": senate_seat["vpi"]["classic"],
            "class": senate_seat["class"]
        }
        senate_seats.append(senate_seat_dict)
        
voter_power_index = pd.DataFrame(senate_seats) 

voter_power_index.head()

len(voter_power_index)

voter_power_index["state"].nunique()

voter_power_index["state"].value_counts()[:5]

total_voter_power_index = (
    voter_power_index
    .groupby("state")
    ["vpi"].sum()
    .to_frame()
)

total_voter_power_index.head()

total_voter_power_index.to_csv(
    "../data/fivethirtyeight/senate-voter-power-index.csv"
)

---

---

---
