==> notebooks/01-analyze-close-races.ipynb <==
# Analyzing Demographics of Close House and Senate Races

import pandas as pd

pd.options.display.max_colwidth = 100

# House Analysis

house_candidates = pd.read_csv("../data/fivethirtyeight/house_candidate_odds.csv")
house_candidates.head()

Calculate FiveThirtyEight's win-probability for each House race's leading candidate:

house_race_details = (
    house_candidates
    .groupby([
        "district",
        "state"
    ])
    .pipe(lambda grp: pd.DataFrame({
        "candidates": grp.size(),
        "max_classic_prob": grp["classic_prob"].max()
    }))
    .reset_index()
)

house_race_details.head()

fips_cross = (
    pd.read_csv(
        "../data/postal-to-fips-crosswalk.csv",
        dtype = {
            "FIPS": str
        }
    )
    .set_index("Code")
)

fips_cross.head()

fips_cross_dict = fips_cross["FIPS"].to_dict()

def make_district_fips(row):
    state_fips = fips_cross_dict[row["state"]]

    # For states with only 1 at-large district, the FIPS is 0, not 1
    if row["state"] in [ "AK", "DE", "MT", "ND", "SD", "VT", "WY" ]:
        district_str = "00"
    else:
        district_str = f"{row['district']:02d}"

    return state_fips + district_str

house_race_details["fips"] = house_race_details.apply(make_district_fips, axis=1)

house_race_details.head()

*Note: `close_races` includes FiveThirtyEight's "Toss-Up" and "Lean" races*

close_races = house_race_details[
    (house_race_details["max_classic_prob"] < 75)
].copy()

other_races = house_race_details[
    (house_race_details["max_classic_prob"] >= 75)
]

print(
    f"There are {len(close_races)} close Houses races, "
    f"and {len(other_races)} other races."
)

## Load and join congressional district demographics

cd_demographics = pd.read_csv("../data/census/CD.csv")

cd_demographics["fips"] = cd_demographics["GEOID"].apply(lambda x: x.split("US")[-1])

cd_demographics.head(5)

cd_totals = (
    cd_demographics[
        cd_demographics["LNTITLE"] == "Total"
    ]
    [[ "fips", "CVAP_EST"]]
    .rename(columns = { "CVAP_EST": "district_total" })
)

assert cd_totals["fips"].nunique() == len(cd_totals)

cd_totals.head()

cd_demographic_pcts = (
    cd_demographics[
        cd_demographics["LNTITLE"] != "Total"
    ]
    [[
        "GEONAME",
        "fips",
        "LNTITLE",
        "CVAP_EST",        
    ]]
    .rename(columns = {
        "LNTITLE": "demographic"
    })
    .merge(
        cd_totals,
        how = "left",
        on = "fips",
        validate = "m:1"
    )
    .assign(
        pct = lambda df: (df["CVAP_EST"] * 100 / df["district_total"]).round(2)
    )
)

cd_demographic_pcts.head()

## Compare close and not close House races

house_comparison = (
    pd.concat([
        house_race_details.assign(category = "all"),
        close_races.assign(category = "close"),
        other_races.assign(category = "other"),
    ])
    .merge(
        cd_demographic_pcts,
        how = "left",
        on = "fips",
    )
    .groupby([
        "demographic",
        "category"
    ])
    [[
        "CVAP_EST",
        "district_total"
    ]]
    .sum()
    .assign(pct = lambda df: (df["CVAP_EST"] * 100 / df["district_total"]).round(3))
    .unstack()
    [[ "CVAP_EST", "pct" ]]
    .sort_values(("pct", "close"), ascending = False)
    .assign(
        close_vs_all = lambda df: df[("pct", "close")] - df[("pct", "all")]
    )
    .reset_index()
)

house_comparison

house_comparison[
    house_comparison["demographic"].str.contains(r"Alone|^Hispanic or Latino$")
]

cd_demographic_pcts[
    (cd_demographic_pcts["demographic"] == "Black or African American Alone") &
    (cd_demographic_pcts["fips"].isin(close_races["fips"])) &
    (cd_demographic_pcts["pct"] <= 2)
][["GEONAME", "demographic", "pct"]].sort_values("pct", ascending = True)

tossup_races = close_races[
    close_races["max_classic_prob"] < 60
].copy()

cd_demographic_pcts[
    (cd_demographic_pcts["demographic"] == "White Alone") &
    (cd_demographic_pcts["fips"].isin(tossup_races["fips"])) &
    (cd_demographic_pcts["pct"] >= 85)
][["GEONAME", "demographic", "pct"]].sort_values("pct", ascending = False)

# Senate analysis

senate_candidates = pd.read_csv("../data/fivethirtyeight/senate_candidate_odds.csv")
senate_candidates.head()

senate_race_details = (
    senate_candidates
    .groupby([
        "state",
        "class",
    ])
    .pipe(lambda grp: pd.DataFrame({
        "candidates": grp.size(),
        "max_classic_prob": grp["classic_prob"].max()
    }))
    .reset_index()
)

senate_race_details["fips"] = senate_race_details["state"].apply(fips_cross_dict.get)

senate_race_details.head()

close_senate_races = senate_race_details[
    (senate_race_details["max_classic_prob"] < 75)
].copy()

other_senate_races = senate_race_details[
    (senate_race_details["max_classic_prob"] >= 75)
]

print(
    f"There are {len(close_senate_races)} close Senate races, "
    f"and {len(other_senate_races)} other races."
)

## Load and join state demographics

state_demographics = pd.read_csv("../data/census/State.csv")
state_demographics["fips"] = state_demographics["GEOID"].apply(lambda x: x.split("US")[-1])
state_demographics.head(5)

state_totals = (
    state_demographics[
        state_demographics["LNTITLE"] == "Total"
    ]
    [[ "fips", "CVAP_EST"]]
    .rename(columns = { "CVAP_EST": "state_total" })
)

assert state_demographics["fips"].nunique() == len(state_totals)

state_totals.head()

state_demographic_pcts = (
    state_demographics[
        state_demographics["LNTITLE"] != "Total"
    ]
    [[
        "GEONAME",
        "fips",
        "LNTITLE",
        "CVAP_EST",        
    ]]
    .rename(columns = {
        "LNTITLE": "demographic"
    })
    .merge(
        state_totals,
        how = "left",
        on = "fips",
        validate = "m:1"
    )
    .assign(
        pct = lambda df: (df["CVAP_EST"] * 100 / df["state_total"]).round(3)
    )
)

state_demographic_pcts.head()

## Compare close and not close Senate races

senate_comparison = (
    pd.concat([
        pd.DataFrame({ "fips": state_totals["fips"].values }).assign(category = "all"),
        close_senate_races[["fips"]].assign(category = "close"),
        other_senate_races[["fips"]].assign(category = "other"),
    ])
    .merge(
        state_demographic_pcts,
        how = "left",
        on = "fips",
    )
    .groupby([
        "demographic",
        "category"
    ])
    [[
        "CVAP_EST",
        "state_total"
    ]]
    .sum()
    .assign(pct = lambda df: (df["CVAP_EST"] * 100 / df["state_total"]).round(3))
    .unstack()
    [[ "CVAP_EST", "pct" ]]
    .sort_values(("pct", "close"), ascending = False)
    .assign(
        close_vs_all = lambda df: df[("pct", "close")] - df[("pct", "all")]
    )
    .reset_index()
)

senate_comparison

senate_comparison[
    senate_comparison["demographic"].str.contains(r"Alone|^Hispanic or Latino$")
]

---

---

---
