==> notebooks/02-analyze-senate-voter-power-index.ipynb <==
# Analyze FiveThirtyEight Voter Power Index

import pandas as pd

## Load "Voter Power Index" scores

voter_power_index = pd.read_csv("../data/fivethirtyeight/senate-voter-power-index.csv")

usps_state = {
    "AZ": "ARIZONA",
    "CA": "CALIFORNIA",
    "CT": "CONNECTICUT",
    "DE": "DELAWARE",
    "FL": "FLORIDA",
    "HI": "HAWAII",
    "IN": "INDIANA",
    "MA": "MASSACHUSETTS",
    "MD": "MARYLAND",
    "ME": "MAINE",
    "MI": "MICHIGAN",
    "MN": "MINNESOTA",
    "MO": "MISSOURI",
    "MS": "MISSISSIPPI",
    "MT": "MONTANA",
    "ND": "NORTH DAKOTA",
    "NE": "NEBRASKA",
    "NJ": "NEW JERSEY",
    "NM": "NEW MEXICO",
    "NV": "NEVADA",
    "NY": "NEW YORK",
    "OH": "OHIO",
    "PA": "PENNSYLVANIA",
    "RI": "RHODE ISLAND",
    "TN": "TENNESSEE",
    "TX": "TEXAS",
    "UT": "UTAH",
    "VA": "VIRGINIA",
    "VT": "VERMONT",
    "WA": "WASHINGTON",
    "WI": "WISCONSIN",
    "WV": "WEST VIRGINIA",
    "WY": "WYOMING"
}

voter_power_index["state_name"] = voter_power_index["state"].apply(usps_state.get)

voter_power_index.head()

assert voter_power_index["state_name"].isnull().sum() == 0

## Load state registered voter counts, by demographic

Note: Population and registered-voter counts are *thousands*.

demo_by_state = pd.read_excel(
    "../data/census/table04b.xls",
    skiprows = 4,
    skipfooter = 5,
    na_values = [ "-" ]
)

demo_by_state.columns = [
   "state", "demographic", "total_pop", "total_citizen_pop", "total_registered", 
   "pct_registered", "moe_registered", "pct_registered_citizen", 
   "moe_registered_citizen", "total_voted", "pct_voted", "moe_voted", 
   "percent_voted_citizen", "moe_voted_citizen"
]

demo_by_state.head()

demo_by_state["state"] = demo_by_state["state"].fillna(method="ffill")

demo_by_state.head()

## Join the two datasets

joined = pd.merge(
    demo_by_state,
    voter_power_index,
    left_on = "state",
    right_on = "state_name",
    how = "left",
    suffixes = ["", "_vpi"]
).dropna(subset = ["vpi"])

joined.head().T

joined["demo_power"] = joined["total_registered"] * joined["vpi"]

trimmed = (
    joined[
        ~joined["demographic"].isin(["Total", "Male", "Female", "White alone"]) &
        ~joined["demographic"].str.contains(r"combination")
    ]
    .sort_values("demo_power", ascending = False)
    .set_index([
        "state",
        "demographic"
    ])
    [[
        "total_registered",
        "vpi",
        "demo_power"
    ]]
)

trimmed.head(20)

*Note: "White non-Hispanic alone" in North Dakota is index 389, while "Black alone or in combination" in Mississippi is index 284*

(
    trimmed.loc[("NORTH DAKOTA", "White non-Hispanic alone")]["demo_power"] 
    / trimmed.loc[("MISSISSIPPI", "Black alone")]["demo_power"]
)

---

---

---
